changes according to changes in `Command` super class
- removed bunch of console.log calls
- corrected other things
- don't mind the interface

+++ bot/commands/DeleteCommand.ts
@@ -2,69 +2,29 @@ import Discord = require('discord.js');
 import readline = require('readline');
 import { Command } from "./Command";
 import { Printer } from '../ui/Printer';
-import { ProgressBar } from '../ui/ProgressBar';
+import { ProgressBar } from '../../ui/web/effects/ProgressBar';
 import { clearTimeout } from 'timers';
 
 export class DeleteCommand extends Command
 {
     private message: Discord.Message;
-    private delete_values: [number, Discord.TextChannel];
+    private delete_values: [number, Discord.TextChannel, string];
 
     public constructor(message: Discord.Message)
     {
-        super("delete", []);
+        super("delete");
         this.message = message;
         this.delete_values = this.getParams(this.parseMessage(message));
     }
 
-    private async overrideDelete(channel: Discord.TextChannel): Promise<string>
-    {
-        let messages: Discord.Collection<string, Discord.Message> = await channel.messages.fetch();
-        let messagesToDelete: Array<Discord.Message> = new Array();
-        messages.forEach(message => { if (message != undefined) messagesToDelete.push(message) });
-        let bar = new ProgressBar(this.delete_values[0], "getting messages");
-        while (messagesToDelete.length < this.delete_values[0]) 
-        {
-            let lastMessageID = await messages.last()?.id;
-            if (lastMessageID == undefined)
-            {
-                console.log(Printer.warn("not more messages to parse, breaking"));
-                break;
-            }
-            else
-            {
-                messages = await channel.messages.fetch({ limit: 50, before: lastMessageID });
-                messages.forEach(message => { if (message != undefined) messagesToDelete.push(message) });
-                readline.moveCursor(process.stdout, 0, -1);
-                readline.clearLine(process.stdout, 0);
-                process.stdout.write("\tnow have " + Printer.info(messagesToDelete.length) + " messages\n");
-            }
-        }
-        bar.start();
-        for (let i = 0; i < messages.size && i < this.delete_values[0]; i++)
-        {
-            let timeout = setTimeout(() =>
-            {
-                readline.moveCursor(process.stdout, 64, -2);
-                console.log(Printer.warn("deleting messages is taking too much time"));
-                readline.moveCursor(process.stdout, 0, 1);
-            }, 100);
-            if (messagesToDelete[i].deletable) await messagesToDelete[i].delete({ timeout: 1 });
-            bar.update(i + 1);
-            clearTimeout(timeout);
-        }
-        console.log("");
-        return "executed";
-    }
-
     public async execute(): Promise<Object> 
     {
-        if (this.delete_values[1] != undefined)
+        console.log(Printer.title("deleting messages"));
+        if (this.delete_values[1] != undefined && this.delete_values[2] == "")
         {
-            console.log(Printer.title("deleting messages"));
             console.log(Printer.args(
-                ["number of messages", "channel name", "method"],
-                [`${this.delete_values[0]}`, `${this.delete_values[1].name}`, "bulk delete"]));
+                ["number of messages", "channel name", "method", "target user"],
+                [`${this.delete_values[0]}`, `${this.delete_values[1].name}`, "bulk delete", `${this.delete_values[2]}`]));
             let channel = this.delete_values[1];
             channel.bulkDelete(this.delete_values[0])
                 .then(response =>
@@ -89,8 +49,107 @@ export class DeleteCommand extends Command
                         .catch(console.error);
                 });
         }
+        else if (this.delete_values[1] != undefined && this.delete_values[2] != "")
+        {
+            console.log(Printer.args(
+                ["number of messages", "channel name", "method", "target user"],
+                [`${this.delete_values[0]}`, `${this.delete_values[1].name}`, "target delete", `${this.delete_values[2]}`]));
+            let channel = this.delete_values[1];
+            this.overrideDelete(channel);
+        }
         else
             return "error";
+    }
+
+    private async overrideDelete(channel: Discord.TextChannel): Promise<string>
+    {
+        let messages: Discord.Collection<string, Discord.Message> = await channel.messages.fetch();
+        if (this.delete_values[2] != "")
+        {
+            messages = messages.filter((message) =>
+            {
+                let username = `${message.author.username.replace(" ", "")}#${message.author.discriminator}`;
+                return username == this.delete_values[2];
+            });
+        }
+        let messagesToDelete: Array<Discord.Message> = new Array();
+        messages.forEach(message => { if (message != undefined) messagesToDelete.push(message) });
+        while (messagesToDelete.length < this.delete_values[0]) 
+        {
+            let lastMessageID = await messages.last()?.id;
+            if (lastMessageID == undefined)
+            {
+                console.log(Printer.warn("not more messages to parse, breaking"));
+                break;
+            }
+            else
+            {
+                messages = await channel.messages.fetch({ limit: 50, before: lastMessageID });
+                if (this.delete_values[2] != "")
+                {
+                    messages = messages.filter((message) =>
+                    {
+                        let username = `${message.author.username.replace(" ", "")}#${message.author.discriminator}`;
+                        return username == this.delete_values[2];
+                    });
+                }
+                messages.forEach(message => { if (message != undefined) messagesToDelete.push(message) });
+            }
+        }
+        let bar = new ProgressBar(this.delete_values[0], "deleting messages");
+        bar.start();
+        for (let i = 0; i < messages.size && i < this.delete_values[0]; i++)
+        {
+            let timeout = setTimeout(() =>
+            {
+                readline.moveCursor(process.stdout, 64, -2);
+                console.log(Printer.warn("deleting messages slower than planned"));
+                readline.moveCursor(process.stdout, 0, 1);
+            }, 10000);
+            if (messagesToDelete[i].deletable) await messagesToDelete[i].delete({ timeout: 1 });
+            bar.update(i + 1);
+            clearTimeout(timeout);
+        }
+        console.log("");
+        return "executed";
+    }
+
+    private getParams(map: Map<string, string>): [number, Discord.TextChannel, string]
+    {
+        let messages = 10;
+        let channel: Discord.TextChannel = undefined;
+        let username = "";
+        if (this.message.channel instanceof Discord.TextChannel) channel = this.message.channel;
+
+        map.forEach((value, key) =>
+        {
+            switch (key)
+            {
+                case "u":
+                    if (/([A-Za-z]+#+[0-9999])\w+/.exec(value)[0] == value)
+                    {
+                        username = value;
+                    }
+                    break;
+                case "n":
+                    if (!Number.isNaN(Number.parseInt(value)))
+                    {
+                        if (!map.has("u"))
+                            messages = Number.parseInt(value) + 1;
+                        else messages = Number.parseInt(value);
+                    }
+                    break;
+                case "c":
+                    let resolvedChannel = this.message.guild.channels.resolve(value);
+                    if (resolvedChannel && resolvedChannel instanceof Discord.TextChannel)
+                    {
+                        channel = resolvedChannel;
+                    }
+                    break;
+                default:
+            }
+        });
+        return [messages, channel, username];
     }
 
     private crop(message: Discord.Message): string
@@ -104,32 +163,9 @@ export class DeleteCommand extends Command
         else res = message.content;
         return res;
     }
+}
 
-    private getParams(args: string[]): [number, Discord.TextChannel]
-    {
-        let messages = 10;
-        let channel: Discord.TextChannel = undefined;
-        if (this.message.channel instanceof Discord.TextChannel)
-            channel = this.message.channel;
-        if (args.length = 2)
-        {
-            // arg 0 -> number of messages
-            if (!Number.isNaN(Number.parseInt(args[0])))
-                messages = Number.parseInt(args[0]);
-            //arg 1 -> channel id
-            let resolvedChannel = this.message.guild.channels.resolve(args[1]);
-            if (resolvedChannel && resolvedChannel instanceof Discord.TextChannel)
-                channel = resolvedChannel;
-        }
-        if (args.length = 1)
-        {
-            // arg 0 -> channel id || number of messages
-            let resolvedChannel = this.message.guild.channels.resolve(args[0]);
-            if (resolvedChannel && resolvedChannel instanceof Discord.TextChannel)
-                channel = resolvedChannel;
-            else if (!Number.isNaN(Number.parseInt(args[0])))
-                messages = Number.parseInt(args[0]);
-        }
-        return [messages, channel];
-    }
+interface Filter
+{
+    sort(value: any, key: any): Function;
 }
